pyproject.toml
#-------------------------------------------------------------------------------
# Project: myst-demo
#-------------------------------------------------------------------------------

[project]

name = "myst-demo"
dynamic = ["version"]
requires-python = ">=3.12"

dependencies = [
    "click>=8.1.7",
    "pyyaml>=6.0.2",
]

#-------------------------------------------------------------------------------
# Build System
#-------------------------------------------------------------------------------

[build-system]

requires = ["hatchling"]
build-backend = "hatchling.build"

#-------------------------------------------------------------------------------
# Build Backend
#-------------------------------------------------------------------------------

[tool.hatch.version]

source = "env"
variable = "PY_VERSION"

#-------------------------------------------------------------------------------
# uv
#-------------------------------------------------------------------------------

[tool.uv]

dev-dependencies = [
    "pytest-cov>=6.0.0",
    "pytest-xdist>=3.6.1",
    "pytest>=8.3.3",
    "python-dotenv>=1.0.1",
    "ruff>=0.7.3",
    "mystmd>=1.3.18",
    "jupyterlab>=4.3.1",
]
